Migrate to an implicit SiloSystem approach
* Disable temporarily staticOnly in the Pickling Protocol since there
  are some issues with it when pickling spores.

* Remove SiloRefFactory since `populate` is now in [[SiloFactory]].
  In the future, we'll readd it to support `fromFun` et al.
