+++ core/src/test/scala/SiloSytemTest.scala
@@ -1,6 +1,9 @@
 package fp
 package test
 
+import fp.backend.SiloSystem
+import fp.backend.netty.SiloSystem
+
 import scala.concurrent.{ Await, ExecutionContext, Future, Promise }
 import scala.concurrent.duration._
 import ExecutionContext.Implicits.global
@@ -11,18 +14,18 @@ import org.scalatest.{ FlatSpec, Matchers }
 class SiloSystemTest extends FlatSpec with Matchers {
 
   "Instantiation" should "yield a default silo system" in {
-    Await.ready(fp.SiloSystem(), 10.seconds) map { _ shouldBe a [fp.SiloSystem] }
+    Await.ready(SiloSystem(), 10.seconds) map { _ shouldBe a [SiloSystem] }
   } 
 
   it should "use the default, Netty-based realization" in {
-    Await.ready(fp.SiloSystem(), 10.seconds) map { _ shouldBe a [fp.backend.SiloSystem] }
+    Await.ready(SiloSystem(), 10.seconds) map { _ shouldBe a [SiloSystem] }
   } 
 
   it should "throw an exception in case of wrong `silo.system.impl` parameter" in {
     val osp = Option(System.getProperty("silo.system.impl"))
 
     System.setProperty("silo.system.impl", "XXX")
-    Await.ready(fp.SiloSystem(), 10.seconds) map { _ shouldBe a [ClassNotFoundException] }
+    Await.ready(SiloSystem(), 10.seconds) map { _ shouldBe a [ClassNotFoundException] }
 
     osp match {
       case None    => System.clearProperty("silo.system.impl")
@@ -31,8 +34,8 @@ class SiloSystemTest extends FlatSpec with Matchers {
   }
 
   it should "throw an exception in case of port is already taken" in {
-    val system = Await.result(fp.SiloSystem(Some(8090)), 10.seconds)
-    Try(Await.result(fp.SiloSystem(Some(8090)), 10.seconds)) match {
+    val system = Await.result(SiloSystem(Some(8090)), 10.seconds)
+    Try(Await.result(SiloSystem(Some(8090)), 10.seconds)) match {
       case Success(s) =>
         Await.result(s.terminate(), 10.seconds)
         true shouldBe false
